+++ src/library/NonTrivialString/definition.ts
