+++ src/library/UniformResourceIdentifier/Data/index.ts
@@ -1,7 +1,7 @@
 import Attempt from '@/library/Attempt';
 import type Base64CharacterEncodedByteSequence from '@/library/Base64CharacterEncodedByteSequence';
 import type ContentDescriptor from '@/library/ContentDescriptor';
-import NonTrivialString from '@/library/NonTrivialString';
+import NonTrivialString from '@/library/NonTrivialString/definition';
 
 import UniformResourceIdentifier from '../index.ts';
